+++ include/handlers.h
@@ -152,6 +152,7 @@ extern int m_wallvoices(struct Client*, struct Client*, int, char*[]);
 extern int m_watch(struct Client*, struct Client*, int, char*[]);
 extern int m_setname(struct Client*, struct Client*, int, char*[]);
 extern int ms_setname(struct Client*, struct Client*, int, char*[]);
+extern int m_tagmsg(struct Client*, struct Client*, int, char*[]);
 extern int m_webirc(struct Client*, struct Client*, int, char*[]);
 extern int m_who(struct Client*, struct Client*, int, char*[]);
 extern int m_whois(struct Client*, struct Client*, int, char*[]);