+++ include/s_user.h
@@ -100,6 +100,7 @@ extern void send_umode(struct Client *cptr, struct Client *sptr,
                        struct Flags *old, int sendset);
 extern void set_snomask(struct Client *, unsigned int, int);
 extern const char *snomask_to_str(unsigned int mask, char *buf, size_t buflen);
+extern unsigned int snomask_str_to_mask(const char *str);
 extern int is_snomask(char *);
 extern int check_target_limit(struct Client *sptr, void *target, const char *name,
     int created);